requirements.txt
autogen-agentchat
autogen-core
autogen-ext[openai,azure]
asyncio
dotenv
openai
psycopg2-binary
streamlit
rich 
pydantic 
